+++ app/render/menus/build.go
@@ -24,65 +24,67 @@ func build(app renderiface.App, world renderiface.GameWorld) {
 	if imgui.BeginMenu("Build") {
 		if imgui.MenuItemBool("Build Navigation Mesh") {
 			fmt.Println("Build Navigation Mesh ")
-			minVertex := mgl64.Vec3{-500, -250, -500}
-			maxVertex := mgl64.Vec3{500, 250, 500}
-			NM = navmesh.New(app, world, minVertex, maxVertex)
-
-			vxs := int(maxVertex.X() - minVertex.X())
-			// vys := int(maxVertex.Y() - minVertex.Y())
-			vzs := int(maxVertex.Z() - minVertex.Z())
-
-			hf := navmesh.NewHeightField(vxs, vzs, minVertex, maxVertex)
-
 			start := time.Now()
-			for _, entity := range world.Entities() {
-				if entity.MeshComponent == nil {
-					continue
-				}
-
-				// triangles := [][3]mgl64.Vec3{
-				// 	{{-50, -50, 50}, {50, -50, 50}, {50, 50, 50}},
-				// 	{{50, 50, 50}, {-50, 50, 50}, {-50, -50, 50}},
-				// 	{{50, 50, -50}, {50, -50, -50}, {-50, -50, -50}},
-				// 	{{-50, -50, -50}, {-50, 50, -50}, {50, 50, -50}},
-				// 	{{50, -50, 50}, {50, -50, -50}, {50, 50, -50}},
-				// 	{{50, 50, -50}, {50, 50, 50}, {50, -50, 50}},
-				// 	{{-50, 50, -50}, {-50, -50, -50}, {-50, -50, 50}},
-				// 	{{-50, -50, 50}, {-50, 50, 50}, {-50, 50, -50}},
-				// 	{{50, 50, 50}, {50, 50, -50}, {-50, 50, 50}},
-				// 	{{-50, 50, 50}, {50, 50, -50}, {-50, 50, -50}},
-				// 	{{-50, -50, 50}, {50, -50, -50}, {50, -50, 50}},
-				// 	{{-50, -50, -50}, {50, -50, -50}, {-50, -50, -50}},
-				// }
-
-				primitives := app.ModelLibrary().GetPrimitives(entity.MeshComponent.MeshHandle)
-				transform := utils.Mat4F64ToF32(entities.WorldTransform(entity))
-				for _, p := range primitives {
-					for i := 0; i < len(p.Primitive.Vertices); i += 3 {
-						v1 := utils.Vec3F32ToF64(transform.Mul4x1(p.Primitive.Vertices[i].Position.Vec4(1)).Vec3())
-						v2 := utils.Vec3F32ToF64(transform.Mul4x1(p.Primitive.Vertices[i+1].Position.Vec4(1)).Vec3())
-						v3 := utils.Vec3F32ToF64(transform.Mul4x1(p.Primitive.Vertices[i+2].Position.Vec4(1)).Vec3())
-
-						navmesh.RasterizeTriangle(
-							int(v1.X()),
-							int(v1.Y()),
-							int(v1.Z()),
-							int(v2.X()),
-							int(v2.Y()),
-							int(v2.Z()),
-							int(v3.X()),
-							int(v3.Y()),
-							int(v3.Z()),
-							hf,
-						)
-					}
-				}
-			}
+			nm := buildNavMesh(app, world)
+			HeightField = nm.HeightField
 			fmt.Println("rasterized voxels in", time.Since(start).Seconds())
-			fmt.Printf("rasterized %d spans\n", hf.SpanCount())
-			HeightField = hf
+			fmt.Printf("rasterized %d spans\n", HeightField.SpanCount())
 		}
 
 		imgui.EndMenu()
 	}
+}
+
+func buildNavMesh(app renderiface.App, world renderiface.GameWorld) *navmesh.NavigationMesh2 {
+	minVertex := mgl64.Vec3{-500, -250, -500}
+	maxVertex := mgl64.Vec3{500, 250, 500}
+	NM = navmesh.New(app, world, minVertex, maxVertex)
+
+	vxs := int(maxVertex.X() - minVertex.X())
+	vzs := int(maxVertex.Z() - minVertex.Z())
+
+	hf := navmesh.NewHeightField(vxs, vzs, minVertex, maxVertex)
+
+	for _, entity := range world.Entities() {
+		if entity.MeshComponent == nil {
+			continue
+		}
+
+		primitives := app.ModelLibrary().GetPrimitives(entity.MeshComponent.MeshHandle)
+		transform := utils.Mat4F64ToF32(entities.WorldTransform(entity))
+		up := mgl64.Vec3{0, 1, 0}
+
+		// rasterize triangles
+		for _, p := range primitives {
+			for i := 0; i < len(p.Primitive.Vertices); i += 3 {
+				v1 := utils.Vec3F32ToF64(transform.Mul4x1(p.Primitive.Vertices[i].Position.Vec4(1)).Vec3())
+				v2 := utils.Vec3F32ToF64(transform.Mul4x1(p.Primitive.Vertices[i+1].Position.Vec4(1)).Vec3())
+				v3 := utils.Vec3F32ToF64(transform.Mul4x1(p.Primitive.Vertices[i+2].Position.Vec4(1)).Vec3())
+
+				tv1 := v2.Sub(v1)
+				tv2 := v3.Sub(v2)
+
+				normal := tv1.Cross(tv2)
+				if normal.LenSqr() > 0 {
+					normal = normal.Normalize()
+				}
+				if normal.Dot(up) > 0.8 {
+					navmesh.RasterizeTriangle(
+						int(v1.X()),
+						int(v1.Y()),
+						int(v1.Z()),
+						int(v2.X()),
+						int(v2.Y()),
+						int(v2.Z()),
+						int(v3.X()),
+						int(v3.Y()),
+						int(v3.Z()),
+						hf,
+					)
+				}
+			}
+		}
+	}
+
+	return &navmesh.NavigationMesh2{HeightField: hf}
 }
