+++ izzet/navmesh/rasterization.go
@@ -1,338 +1,556 @@
 package navmesh
 
-import (
-	"fmt"
-	"math"
-	"sort"
-	"sync"
-	"time"
+import "math"
 
-	"github.com/go-gl/mathgl/mgl64"
-	"github.com/kkevinchou/izzet/app/entities"
-	"github.com/kkevinchou/kitolib/collision"
-	"github.com/kkevinchou/kitolib/collision/collider"
-	"github.com/kkevinchou/kitolib/utils"
-)
+var offset int = 100
 
-type MinXTriangles []Triangle
+func Plinex(x0, y0, z0, x1, y1, z1 int, LY, LZ, RY, RZ []int, vxs int) {
+	var i, n, cx, cy, cz, sx, sy, sz int
+	var by, bz []int
 
-func (u MinXTriangles) Len() int {
-	return len(u)
-}
-func (u MinXTriangles) Swap(i, j int) {
-	u[i], u[j] = u[j], u[i]
-}
-func (u MinXTriangles) Less(i, j int) bool {
-	return u[i].MinX < u[j].MinX
-}
+	// target buffer & order points by x
+	if x1 >= x0 {
+		i = x0
+		x0 = x1
+		x1 = i
 
-// v1, v2, v3 := triangle.Vertices[0], triangle.Vertices[1], triangle.Vertices[2]
+		i = y0
+		y0 = y1
+		y1 = i
 
-// minX := math.Min(math.Min(v1.X(), v2.X()), v3.X())
-// maxX := math.Max(math.Max(v1.X(), v2.X()), v3.X())
-// minY := math.Min(math.Min(v1.Y(), v2.Y()), v3.Y())
-// maxY := math.Max(math.Max(v1.Y(), v2.Y()), v3.Y())
-// minZ := math.Min(math.Min(v1.Z(), v2.Z()), v3.Z())
-// maxZ := math.Max(math.Max(v1.Z(), v2.Z()), v3.Z())
+		i = z0
+		z0 = z1
+		z1 = i
 
-func (n *NavigationMesh) Voxelize2(triangles []Triangle2) {
-	var voxels []mgl64.Vec3
-	for _, triangle := range triangles {
-		for i := range 3 {
-			v1 := triangle.Vertices[i]
-			v2 := triangle.Vertices[(i+1)%3]
-			n.DebugLines = append(n.DebugLines, [2]mgl64.Vec3{v1, v2})
+		by = LY
+		bz = LZ
+	} else {
+		by = RY
+		bz = RZ
+	}
 
-			dx := v2.X() - v1.X()
-			dy := v2.Y() - v1.Y()
-			dz := v2.Z() - v1.Z()
+	// line DDA parameters
+	x1 -= x0
+	sx = 0
+	if x1 > 0 {
+		sx = 1
+	}
+	if x1 < 0 {
+		sx = -1
+		x1 = -x1
+	}
+	if x1 != 0 {
+		x1++
+	}
+	n = x1
 
-			steps := int(math.Abs(dx))
-			if math.Abs(dy) > math.Abs(dx) && math.Abs(dy) > math.Abs(dz) {
-				steps = int(math.Abs(dy))
-			} else if math.Abs(dz) > math.Abs(dx) && math.Abs(dz) > math.Abs(dy) {
-				steps = int(math.Abs(dz))
-			}
+	y1 -= y0
+	sy = 0
+	if y1 > 0 {
+		sy = 1
+	}
+	if y1 < 0 {
+		sy = -1
+		y1 = -y1
+	}
+	if y1 != 0 {
+		y1++
+	}
+	if n < y1 {
+		n = y1
+	}
 
-			xInc := float64(dx) / float64(steps)
-			yInc := float64(dy) / float64(steps)
-			zInc := float64(dz) / float64(steps)
+	z1 -= z0
+	sz = 0
+	if z1 > 0 {
+		sz = 1
+	}
+	if z1 < 0 {
+		sz = -1
+		z1 = -z1
+	}
+	if z1 != 0 {
+		z1++
+	}
+	if n < z1 {
+		n = z1
+	}
 
-			currentX := v1.X()
-			currentY := v1.Y()
-			currentZ := v1.Z()
-
-			for _ = range steps + 1 {
-				renderX := math.Round(currentX)
-				renderY := math.Round(currentY)
-				renderZ := math.Round(currentZ)
-				// _, fracY := math.Modf(currentY)
-				// renderY := math.Round(currentY)
-				// if fracY >= 0.5 {
-				// 	renderY = math.Ceil(currentY)
-				// }
-
-				// _, fracX := math.Modf(currentX)
-				// renderX := math.Floor(currentX)
-				// if math.Abs(fracX) >= 0.5 {
-				// 	renderX = math.Ceil(currentX)
-				// }
-
-				// _, fracZ := math.Modf(currentZ)
-				// renderZ := math.Floor(currentZ)
-				// if fracZ >= 0.5 {
-				// 	renderZ = math.Ceil(currentZ)
-				// }
-
-				voxels = append(voxels, mgl64.Vec3{renderX, renderY, renderZ})
-
-				currentX += xInc
-				currentY += yInc
-				currentZ += zInc
-			}
+	// single pixel (not a line)
+	if n == 0 {
+		if x0 >= 0 && x0 < vxs {
+			LY[x0] = y0
+			LZ[x0] = z0
+			RY[x0] = y0
+			RZ[x0] = z0
 		}
+		return
 	}
 
-	n.DebugVoxels = voxels
-}
-
-// TODO: this method is fairly brute force and could be improved a lot, Recast does this more efficiently.
-// the current implementation considers all voxel positions in a region and does an intersection test with
-// geometry. This is quite wasteful as a lot of geometry will not be intersecting with the voxel but we will
-// consider it anyway. Spatial partitioning helps reduce the number of checks but still isn't ideal. This is
-// akin to raytracing in graphical rendering. Instead we should take the rasterization approach (ironically what
-// this file is named), which in starts with taking the geometry and finding all voxels that it intersects with.
-// Starting from the geometry allows us to quickly fill in a large amount of voxels with minimal intersection
-// checks which is the expensive part of voxelization.
-func (n *NavigationMesh) voxelize() [][][]Voxel {
-	start := time.Now()
-	spatialPartition := n.world.SpatialPartition()
-	sEntities := spatialPartition.QueryEntities(n.Volume)
-
-	var candidateEntities []*entities.Entity
-	boundingBoxes := map[int]collider.BoundingBox{}
-	// entityTriangles := map[int][]Triangle{}
-	entityTriCount := map[int]int{}
-	var triangles []Triangle
-
-	// collect candidate entities for voxelization
-	for _, entity := range sEntities {
-		e := n.world.GetEntityByID(entity.GetID())
-		if e == nil {
-			continue
+	// ND DDA algo i is parameter
+	for cx, cy, cz, i = n, n, n, 0; i < n; i++ {
+		if x0 >= 0 && x0 < vxs {
+			by[x0] = y0
+			bz[x0] = z0
 		}
-
-		if entity.GetID() != 541 {
-			continue
+		cx -= x1
+		if cx <= 0 {
+			cx += n
+			x0 += sx
 		}
-
-		candidateEntities = append(candidateEntities, e)
-		boundingBoxes[e.GetID()] = e.BoundingBox()
-		transform := utils.Mat4F64ToF32(entities.WorldTransform(e))
-
-		primitives := n.app.ModelLibrary().GetPrimitives(e.MeshComponent.MeshHandle)
-
-		// for _, rd := range e.Model.RenderData() {
-		for _, mlPrimitive := range primitives {
-			primitive := mlPrimitive.Primitive
-			for i := 0; i < len(primitive.Vertices); i += 3 {
-				t := Triangle{
-					V1: utils.Vec3F32ToF64(transform.Mul4x1(primitive.Vertices[i].Position.Vec4(1)).Vec3()),
-					V2: utils.Vec3F32ToF64(transform.Mul4x1(primitive.Vertices[i+1].Position.Vec4(1)).Vec3()),
-					V3: utils.Vec3F32ToF64(transform.Mul4x1(primitive.Vertices[i+2].Position.Vec4(1)).Vec3()),
-				}
-				t.MinX = math.Min(t.V1.X(), math.Min(t.V2.X(), t.V3.X()))
-				triangles = append(triangles, t)
-			}
-			numVerts := len(primitive.Vertices)
-			entityTriCount[e.GetID()] = numVerts / 3
+		cy -= y1
+		if cy <= 0 {
+			cy += n
+			y0 += sy
 		}
-	}
-
-	// for more complex geometry it may be worth actually creating an
-	// oct tree for the mesh. realistically we shouldn't be using
-	// very complicated geometry for generating nav meshes
-	sort.Sort(MinXTriangles(triangles))
-
-	outputWork := make(chan OutputWork)
-	delta := n.Volume.MaxVertex.Sub(n.Volume.MinVertex)
-	var dimensions [3]int = [3]int{int(delta[0] / n.voxelDimension), int(delta[1] / n.voxelDimension), int(delta[2] / n.voxelDimension)}
-
-	inputWorkCount := dimensions[0] * dimensions[1] * dimensions[2]
-	inputWork := make(chan VoxelPosition, inputWorkCount)
-	workerCount := 12
-
-	doneWorkerCount := 0
-	var doneWorkerMutex sync.Mutex
-
-	// set up workers perform voxelization at a specific 3d coordinate
-	for i := 0; i < workerCount; i++ {
-		go func() {
-			for input := range inputWork {
-				x, y, z := input[0], input[1], input[2]
-
-				voxel := &collider.BoundingBox{
-					MinVertex: n.Volume.MinVertex.Add(mgl64.Vec3{float64(x), float64(y), float64(z)}.Mul(n.voxelDimension)),
-					MaxVertex: n.Volume.MinVertex.Add(mgl64.Vec3{float64(x + 1), float64(y + 1), float64(z + 1)}.Mul(n.voxelDimension)),
-				}
-				voxelAABB := AABB{Min: voxel.MinVertex, Max: voxel.MaxVertex}
-
-				for _, entity := range candidateEntities {
-					bb := boundingBoxes[entity.GetID()]
-					if !collision.CheckOverlapAABBAABB(voxel, &bb) {
-						continue
-					}
-
-					for _, tri := range triangles {
-						// NOTE - rather than doing an expensive AABB/Triangle intersection
-						// Recast clips the triangle against the voxels in the heighfield.
-						// that implementation is likely a lot more performant
-						if IntersectAABBTriangle(voxelAABB, tri) {
-							outputWork <- OutputWork{
-								x:           x,
-								y:           y,
-								z:           z,
-								boundingBox: *voxel,
-							}
-
-							goto Done
-						}
-						if voxelAABB.Max.X() < tri.MinX {
-							continue
-						}
-					}
-				Done:
-				}
-			}
-
-			doneWorkerMutex.Lock()
-			doneWorkerCount++
-			if doneWorkerCount == workerCount {
-				fmt.Println("generation time seconds", time.Since(start).Seconds())
-				close(outputWork)
-			}
-			doneWorkerMutex.Unlock()
-		}()
-	}
-
-	// initialize the voxel field
-	voxelField := make([][][]Voxel, dimensions[0])
-	for i := range voxelField {
-		voxelField[i] = make([][]Voxel, dimensions[1])
-		for j := range voxelField[i] {
-			voxelField[i][j] = make([]Voxel, dimensions[2])
-		}
-	}
-
-	// create a work item for each voxel location. this is the brute force approach which is not very efficient
-	for i := 0; i < dimensions[0]; i++ {
-		for j := 0; j < dimensions[1]; j++ {
-			for k := 0; k < dimensions[2]; k++ {
-				inputWork <- VoxelPosition{i, j, k}
-			}
-		}
-	}
-	close(inputWork)
-
-	// assemble voxels into the voxel field
-	for work := range outputWork {
-		n.voxelCount++
-
-		x, y, z := work.x, work.y, work.z
-		voxelField[x][y][z] = NewVoxel(x, y, z)
-		voxelField[x][y][z].Filled = true
-	}
-	fmt.Printf("generated %d voxels\n", n.voxelCount)
-
-	totalTricount := 0
-	for _, count := range entityTriCount {
-		totalTricount += count
-	}
-	fmt.Println("nav mesh entity tri count", totalTricount)
-	return voxelField
-}
-
-func buildNavigableArea(voxelField [][][]Voxel, dimensions [3]int) {
-	work := make(chan [2]int, dimensions[0]*dimensions[2])
-	workerCount := 12
-
-	var wg sync.WaitGroup
-	wg.Add(workerCount)
-
-	// remove voxels that are too low - i.e. there is a voxel from above that
-	// would interfere with the agent height
-	for wc := 0; wc < workerCount; wc++ {
-		go func() {
-			defer wg.Done()
-			for w := range work {
-				for y := dimensions[1] - 1; y >= 0; y-- {
-					x, z := w[0], w[1]
-					if voxelField[x][y][z].Filled {
-						for i := 1; i < agentHeight+1; i++ {
-							if y-i < 0 {
-								break
-							}
-							voxelField[x][y-i][z] = NewVoxel(x, y-i, z)
-						}
-						y -= agentHeight
-					}
-				}
-			}
-		}()
-	}
-
-	for i := 0; i < dimensions[0]; i++ {
-		for j := 0; j < dimensions[2]; j++ {
-			work <- [2]int{i, j}
-		}
-	}
-
-	close(work)
-	wg.Wait()
-}
-
-func fillHoles(voxelField [][][]Voxel, dimensions [3]int) {
-	for y := 0; y < dimensions[1]; y++ {
-		for z := 0; z < dimensions[2]; z++ {
-			for x := 0; x < dimensions[0]; x++ {
-				if voxelField[x][y][z].Filled {
-					continue
-				}
-
-				// voxel := &voxelField[x][y][z]
-			}
+		cz -= z1
+		if cz <= 0 {
+			cz += n
+			z0 += sz
 		}
 	}
 }
 
-type Voxel2 struct {
-	Filled  bool
-	X, Y, Z int
-}
+func Pliney(x0, y0, z0, x1, y1, z1 int, LX, LZ, RX, RZ []int, vys int) {
+	var i, n, cx, cy, cz, sx, sy, sz int
+	var bx, bz []int
 
-type Voxel struct {
-	Filled           bool
-	X, Y, Z          int
-	DistanceField    float64
-	Seed             bool
-	RegionID         int
-	DEBUGCOLORFACTOR *float32
-	Border           bool
-	ContourCorner    bool
-}
+	// target buffer & order points by y
+	if y1 >= y0 {
+		i = x0
+		x0 = x1
+		x1 = i
 
-type OutputWork struct {
-	x, y, z     int
-	boundingBox collider.BoundingBox
-}
+		i = y0
+		y0 = y1
+		y1 = i
 
-func NewVoxel(x, y, z int) Voxel {
-	return Voxel{
-		Filled:        false,
-		X:             x,
-		Y:             y,
-		Z:             z,
-		DistanceField: MaxDistanceFieldValue,
-		RegionID:      -1,
+		i = z0
+		z0 = z1
+		z1 = i
+
+		bx = LX
+		bz = LZ
+	} else {
+		bx = RX
+		bz = RZ
 	}
+
+	// line DDA parameters
+	x1 -= x0
+	sx = 0
+	if x1 > 0 {
+		sx = 1
+	}
+	if x1 < 0 {
+		sx = -1
+		x1 = -x1
+	}
+	if x1 != 0 {
+		x1++
+	}
+	n = x1
+
+	y1 -= y0
+	sy = 0
+	if y1 > 0 {
+		sy = 1
+	}
+	if y1 < 0 {
+		sy = -1
+		y1 = -y1
+	}
+	if y1 != 0 {
+		y1++
+	}
+	if n < y1 {
+		n = y1
+	}
+
+	z1 -= z0
+	sz = 0
+	if z1 > 0 {
+		sz = 1
+	}
+	if z1 < 0 {
+		sz = -1
+		z1 = -z1
+	}
+	if z1 != 0 {
+		z1++
+	}
+	if n < z1 {
+		n = z1
+	}
+
+	// single pixel (not a line)
+	if n == 0 {
+		if y0 >= 0 && y0 < vys {
+			LX[y0] = x0
+			LZ[y0] = z0
+			RX[y0] = x0
+			RZ[y0] = z0
+		}
+		return
+	}
+
+	// ND DDA algo i is parameter
+	for cx, cy, cz, i = n, n, n, 0; i < n; i++ {
+		if y0 >= 0 && y0 < vys {
+			bx[y0] = x0
+			bz[y0] = z0
+		}
+		cx -= x1
+		if cx <= 0 {
+			cx += n
+			x0 += sx
+		}
+		cy -= y1
+		if cy <= 0 {
+			cy += n
+			y0 += sy
+		}
+		cz -= z1
+		if cz <= 0 {
+			cz += n
+			z0 += sz
+		}
+	}
+}
+
+func Plinez(x0, y0, z0, x1, y1, z1 int, LX, LY, RX, RY []int, vzs int) {
+	var i, n, cx, cy, cz, sx, sy, sz int
+	var bx, by []int
+
+	// target buffer & order points by z
+	if z1 >= z0 {
+		i = x0
+		x0 = x1
+		x1 = i
+
+		i = y0
+		y0 = y1
+		y1 = i
+
+		i = z0
+		z0 = z1
+		z1 = i
+
+		bx = LX
+		by = LY
+	} else {
+		bx = RX
+		by = RY
+	}
+
+	// line DDA parameters
+	x1 -= x0
+	sx = 0
+	if x1 > 0 {
+		sx = 1
+	}
+	if x1 < 0 {
+		sx = -1
+		x1 = -x1
+	}
+	if x1 != 0 {
+		x1++
+	}
+	n = x1
+
+	y1 -= y0
+	sy = 0
+	if y1 > 0 {
+		sy = 1
+	}
+	if y1 < 0 {
+		sy = -1
+		y1 = -y1
+	}
+	if y1 != 0 {
+		y1++
+	}
+	if n < y1 {
+		n = y1
+	}
+
+	z1 -= z0
+	sz = 0
+	if z1 > 0 {
+		sz = 1
+	}
+	if z1 < 0 {
+		sz = -1
+		z1 = -z1
+	}
+	if z1 != 0 {
+		z1++
+	}
+	if n < z1 {
+		n = z1
+	}
+
+	// single pixel (not a line)
+	if n == 0 {
+		if z0 >= 0 && z0 < vzs {
+			LX[z0] = x0
+			LY[z0] = y0
+			RX[z0] = x0
+			RY[z0] = y0
+		}
+		return
+	}
+
+	// ND DDA algo i is parameter
+	for cx, cy, cz, i = n, n, n, 0; i < n; i++ {
+		if z0 >= 0 && z0 < vzs {
+			bx[z0] = x0
+			by[z0] = y0
+		}
+		cx -= x1
+		if cx <= 0 {
+			cx += n
+			x0 += sx
+		}
+		cy -= y1
+		if cy <= 0 {
+			cy += n
+			y0 += sy
+		}
+		cz -= z1
+		if cz <= 0 {
+			cz += n
+			z0 += sz
+		}
+	}
+}
+
+func Line(x0, y0, z0, x1, y1, z1 int, c float32, vxs, vys, vzs int, heightField *HeightField) int {
+	var i, n, cx, cy, cz, sx, sy, sz int
+
+	// line DDA parameters
+	x1 -= x0
+	sx = 0
+	if x1 > 0 {
+		sx = 1
+	}
+	if x1 < 0 {
+		sx = -1
+		x1 = -x1
+	}
+	if x1 != 0 {
+		x1++
+	}
+	n = x1
+
+	y1 -= y0
+	sy = 0
+	if y1 > 0 {
+		sy = 1
+	}
+	if y1 < 0 {
+		sy = -1
+		y1 = -y1
+	}
+	if y1 != 0 {
+		y1++
+	}
+	if n < y1 {
+		n = y1
+	}
+
+	z1 -= z0
+	sz = 0
+	if z1 > 0 {
+		sz = 1
+	}
+	if z1 < 0 {
+		sz = -1
+		z1 = -z1
+	}
+	if z1 != 0 {
+		z1++
+	}
+	if n < z1 {
+		n = z1
+	}
+
+	count := 0
+
+	// single pixel (not a line)
+	if n == 0 {
+		if x0 >= 0 && x0 < vxs && y0 >= 0 && y0 < vys && z0 >= 0 && z0 < vzs {
+			heightField.AddVoxel(x0, y0, z0)
+		}
+		return count
+	}
+
+	// ND DDA algo i is parameter
+	for cx, cy, cz, i = n, n, n, 0; i < n; i++ {
+		if x0 >= 0 && x0 < vxs && y0 >= 0 && y0 < vys && z0 >= 0 && z0 < vzs {
+			heightField.AddVoxel(x0, y0, z0)
+		}
+		cx -= x1
+		if cx <= 0 {
+			cx += n
+			x0 += sx
+		}
+		cy -= y1
+		if cy <= 0 {
+			cy += n
+			y0 += sy
+		}
+		cz -= z1
+		if cz <= 0 {
+			cz += n
+			z0 += sz
+		}
+	}
+
+	return count
+}
+
+func RasterizeTriangle(x0, y0, z0, x1, y1, z1, x2, y2, z2 int, heightField *HeightField) int {
+	vxs := int(heightField.bMax.X() - heightField.bMin.X())
+	vys := int(heightField.bMax.Y() - heightField.bMin.Y())
+	vzs := int(heightField.bMax.Z() - heightField.bMin.Z())
+
+	vsz := int(math.Max(math.Max(float64(vxs), float64(vys)), float64(vzs)))
+
+	LX := make([]int, vsz)
+	LY := make([]int, vsz)
+	LZ := make([]int, vsz)
+	RX := make([]int, vsz)
+	RY := make([]int, vsz)
+	RZ := make([]int, vsz)
+
+	var X0, Y0, Z0, X1, Y1, Z1, dx, dy, dz, x, y, z int
+
+	bMin := heightField.BMin()
+
+	x0 -= int(bMin.X())
+	x1 -= int(bMin.X())
+	x2 -= int(bMin.X())
+
+	y0 -= int(bMin.Y())
+	y1 -= int(bMin.Y())
+	y2 -= int(bMin.Y())
+
+	z0 -= int(bMin.Z())
+	z1 -= int(bMin.Z())
+	z2 -= int(bMin.Z())
+
+	// BBOX
+	X0 = x0
+	X1 = x0
+	if X0 > x1 {
+		X0 = x1
+	}
+	if X1 < x1 {
+		X1 = x1
+	}
+	if X0 > x2 {
+		X0 = x2
+	}
+	if X1 < x2 {
+		X1 = x2
+	}
+	Y0 = y0
+	Y1 = y0
+	if Y0 > y1 {
+		Y0 = y1
+	}
+	if Y1 < y1 {
+		Y1 = y1
+	}
+	if Y0 > y2 {
+		Y0 = y2
+	}
+	if Y1 < y2 {
+		Y1 = y2
+	}
+	Z0 = z0
+	Z1 = z0
+	if Z0 > z1 {
+		Z0 = z1
+	}
+	if Z1 < z1 {
+		Z1 = z1
+	}
+	if Z0 > z2 {
+		Z0 = z2
+	}
+	if Z1 < z2 {
+		Z1 = z2
+	}
+
+	dx = X1 - X0
+	dy = Y1 - Y0
+	dz = Z1 - Z0
+
+	var count int
+
+	if dx >= dy && dx >= dz { // x is major axis
+		// render circumference into left/right buffers
+		Plinex(x0, y0, z0, x1, y1, z1, LY, LZ, RY, RZ, vxs)
+		Plinex(x1, y1, z1, x2, y2, z2, LY, LZ, RY, RZ, vxs)
+		Plinex(x2, y2, z2, x0, y0, z0, LY, LZ, RY, RZ, vxs)
+
+		// fill the triangle
+		if X0 < 0 {
+			X0 = 0
+		}
+		if X1 >= vxs {
+			X1 = vxs - 1
+		}
+
+		for x = X0; x <= X1; x++ {
+			y0 = LY[x]
+			z0 = LZ[x]
+			y1 = RY[x]
+			z1 = RZ[x]
+			count += Line(x, y0, z0, x, y1, z1, 1, vxs, vys, vzs, heightField)
+		}
+	} else if dy >= dx && dy >= dz { // y is major axis
+		// render circumference into left/right buffers
+		Pliney(x0, y0, z0, x1, y1, z1, LX, LZ, RX, RZ, vys)
+		Pliney(x1, y1, z1, x2, y2, z2, LX, LZ, RX, RZ, vys)
+		Pliney(x2, y2, z2, x0, y0, z0, LX, LZ, RX, RZ, vys)
+
+		// fill the triangle
+		if Y0 < 0 {
+			Y0 = 0
+		}
+		if Y1 >= vys {
+			Y1 = vys - 1
+		}
+		for y = Y0; y <= Y1; y++ {
+			x0 = LX[y]
+			z0 = LZ[y]
+			x1 = RX[y]
+			z1 = RZ[y]
+			count += Line(x0, y, z0, x1, y, z1, 1, vxs, vys, vzs, heightField)
+		}
+	} else if dz >= dx && dz >= dy { // z is major axis
+		// render circumference into left/right buffers
+		Plinez(x0, y0, z0, x1, y1, z1, LX, LY, RX, RY, vzs)
+		Plinez(x1, y1, z1, x2, y2, z2, LX, LY, RX, RY, vzs)
+		Plinez(x2, y2, z2, x0, y0, z0, LX, LY, RX, RY, vzs)
+
+		// fill the triangle
+		if Z0 < 0 {
+			Z0 = 0
+		}
+		if Z1 >= vzs {
+			Z1 = vzs - 1
+		}
+		for z = Z0; z <= Z1; z++ {
+			x0 = LX[z]
+			y0 = LY[z]
+			x1 = RX[z]
+			y1 = RY[z]
+			count += Line(x0, y0, z, x1, y1, z, 1, vxs, vys, vzs, heightField)
+		}
+	}
+	return count
 }
